requirements.txt
aiogram==2.16
aiohttp==3.8.0
aiohttp-jinja2==1.5
aiosignal==1.2.0
async-timeout==4.0.0
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
colorama==0.4.4
frozenlist==1.2.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
numpy==1.21.4
Pillow==8.4.0
pytz==2021.3
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchvision==0.11.1+cu113
tqdm==4.62.3
typing-extensions==3.10.0.2
yarl==1.7.2
